Move spectral line files to subfolder
