+++ SpectralLine/SpectralLineCrawler.h
